+++ src/main/kotlin/com/atlassian/performance/tools/aws/api/ami/tiebreaker/NewestAvailableAmiTiebreaker.kt
@@ -1,0 +1,14 @@
+package com.atlassian.performance.tools.aws.api.ami.tiebreaker
+
+import com.amazonaws.services.ec2.model.Image
+import com.amazonaws.services.ec2.model.ImageState
+
+class NewestAvailableAmiTiebreaker : AmiTiebreaker {
+    override fun pick(amis: List<Image>): Image? {
+        return amis
+            .asSequence()
+            .filter { ImageState.fromValue(it.state) == ImageState.Available }
+            .sortedByDescending { it.creationDate }
+            .firstOrNull()
+    }
+}