+++ src/test/kotlin/com/atlassian/performance/tools/aws/api/ami/SshAmiModIT.kt
@@ -9,6 +9,7 @@ import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import java.nio.file.Files.createTempDirectory
 import java.time.Duration
+import java.util.*
 
 class SshAmiModIT {
 
@@ -25,7 +26,9 @@ class SshAmiModIT {
                 "echo-file" to "some-file.txt",
             )
         }
-        val sshAmiMod = SshAmiMod.Builder(echo).build()
+        val sshAmiMod = SshAmiMod.Builder(echo)
+            .amiCache(NoAmiCache())
+            .build()
 
         // when
         val newImageId = sshAmiMod.provideAmiId(aws)
@@ -40,5 +43,34 @@ class SshAmiModIT {
             val actualContent = ssh.safeExecute("cat some-file.txt").output
             assertThat(actualContent).isEqualToIgnoringWhitespace("kebab")
         }
+    }
+
+    @Test
+    fun shouldReuseCachedAmi() {
+        // given
+        val uniqueTags = mapOf(
+            "cache-key1" to UUID.randomUUID().toString(),
+            "cache-key2" to UUID.randomUUID().toString(),
+        )
+        val brandNewMod = object : SshInstanceMod {
+            var amisCreated = 0
+
+            override fun modify(sshInstance: SshInstance) {
+                amisCreated++
+            }
+
+            override fun tag() = uniqueTags
+        }
+        val sshAmiMod = SshAmiMod.Builder(brandNewMod)
+            .amiCache(TiebreakingAmiCache.Builder().build())
+            .build()
+
+        // when
+        val firstImageId = sshAmiMod.provideAmiId(aws)
+        val secondImageId = sshAmiMod.provideAmiId(aws)
+
+        // then
+        assertThat(secondImageId).isEqualTo(firstImageId)
+        assertThat(brandNewMod.amisCreated).isEqualTo(1)
     }
 }