+++ CHANGELOG.md
@@ -27,6 +27,7 @@ Dropping a requirement of a major version of a dependency is a new contract.
 - Inject `AmiProvider` into `Aws` to choose the default AMI.
 - Expose `CanonicalAmiProvider` to find Ubuntu AMIs by name.
 - Add `SshAmiMod` to modify existing AMIs.
+- Add `AmiCache` to `SshAmiMod` to reuse AMIs instead of recreating them every time.
 
 ## [1.9.1] - 2023-01-24
 [1.9.1]: https://bitbucket.org/atlassian/aws-resources/branches/compare/release-1.9.1%0Drelease-1.9.0